refactor test and IsInf; fix format
