+++ src/indexes/scoring/scorer.cc
@@ -7,9 +7,20 @@
 
 #include "src/indexes/scoring/scorer.h"
 
+#include <cstdint>
+#include <cstring>
+
 #include "absl/types/span.h"
 
 namespace valkey_search::indexes::scoring {
+
+bool IsInf(float f) {
+  static constexpr uint32_t kExponentMask = 0x7F800000U;
+  static constexpr uint32_t kMantissaMask = 0x007FFFFFU;
+  uint32_t bits;
+  std::memcpy(&bits, &f, sizeof(bits));
+  return (bits & kExponentMask) == kExponentMask && (bits & kMantissaMask) == 0;
+}
 
 float Scorer::CombineGroup(absl::Span<const float> child_scores,
                            float group_weight) const {
